+++ src/graphql/mutation/index.ts
@@ -1,0 +1,11 @@
+import * as WorkbookMutations from "./workbook"
+import * as SimMutations from "./sim"
+import * as UserMutations from "./user"
+
+const Mutation = {
+  ...WorkbookMutations,
+  ...SimMutations,
+  ...UserMutations
+};
+
+export default Mutation;
