+++ src/graphql/query/index.ts
@@ -1,0 +1,9 @@
+import * as SimQueries from "./sim";
+import * as WorkbookQueries from "./workbook";
+
+const Sim = {
+  ...SimQueries,
+  ...WorkbookQueries,
+};
+
+export default Sim;
