+++ src/graphql/typeDefs/user.ts
@@ -1,0 +1,14 @@
+export const UserTypes = `
+  type User {
+    _id: ID!
+    email: String!
+    password: String
+  }
+
+  input UserInput {
+    email: String!
+    password: String!
+  }
+`
+
+export default UserTypes
